backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
groq==0.4.1
pydub==0.25.1
numpy==1.24.3
PyPDF2==3.0.1
aiofiles==23.2.1
python-multipart==0.0.6